pre-commit: Add exemplar's checks, make them pass, add them to CI
Previously, although infra provided beman-tidy as a pre-commit hook to
consumers of the repository, it did not use pre-commit checks on its
own source.

This commit addresses this by adding all of the same pre-commit checks
as are used in exemplar, except for clang-format and CMake linting,
since those aren't relevant here.

It fixes all of the failed checks and adds pre-commit checking to CI
so they don't recur.