Implement allow_untracked_files
This parameter allows beman submodules to be installed into a
subdirectory without being the exclusive occupant of that
subdirectory; it can contain other files which the beman submodule
will clobber if they conflict with the submodule's repository, but it
will not automatically remove those files.

It is not enabled by default because it carries the risk that a file
that was removed from the beman submodule's upstream will not be
removed from the subdirectory.