Polish beman_install_library
This commit:

* Changes defaults such that one must opt out of installing
CMake config packages for beman libraries using
beman_install_library. This is because if one is installing a
beman library, it is more likely than not that they would want
to support potential CMake consumers using those libraries.

* Fixes a bug that caused the `<NAME>__INSTALL_CONFIG_FILE_PACKAGE`
option to not work as designed. This is because of incorrect
quoting in the relevant CMake conditional statement in the
implementation of `beman_install_library`.

* Adds detail the the `README.md` such that the existence of the
`cmake/` subdirectory is documented. API documentation for
`beman_install_library` was also added.